% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame.boot.R
\name{ame.boot}
\alias{ame.boot}
\title{Bootrapping zur Bestimmung  der Standardfehler und 95\%-Konfidenzintervalle}
\usage{
ame.boot(data.name, rep = 100, meth = "dt", func, var.name, fromtoby)
}
\description{
Bootrapping zur Bestimmung  der Standardfehler und 95\%-Konfidenzintervalle
}
\examples{
\dontrun{
ame.boot(test.data, 10, "lm", y ~ alter + geschl + bildung, "geschl")
ame.boot(test.data, 10, "dtt", y ~ alter+ bildung, "geschl")
ame.boot(test.data, 10, "dt", y ~ alter + geschl + bildung, "bildung")
ame.boot(test.data, 10, "dt", y ~ alter + geschl, "alter", seq(20,60,1))
}
}

