\name{mlame-package}
\alias{mlame-package}
\alias{mlame}
\docType{package}
\title{
\packageTitle{mlame}
}
\description{
\packageDescription{mlame}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mlame}
\packageIndices{mlame}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mlame}

Maintainer: \packageMaintainer{mlame}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
