% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlame.R
\docType{package}
\name{mlame}
\alias{mlame}
\alias{mlame-package}
\title{mlame: Estimate average marginal effects for machine learning algorithms.}
\description{
An easy way to estimate average marginal effects (AME), predicted
probabilities / values and similar quantities from different kinds of
machine learning algorithms. Bootstrap estimates of standard errors,
confidence intervals and p-values are also provided.
}
\section{mlame functions}{

So far mlame provides two functions \code{ame} and \code{ame.boot}.
}

