% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ame.R
\name{ame}
\alias{ame}
\title{Average marginal effects fuer metrische abhaengige Variablen}
\usage{
ame(data.name, meth = "dt", func, var.name, fromtoby = NULL,
  plotTree = FALSE, plotPV = FALSE)
}
\description{
Average marginal effects fuer metrische abhaengige Variablen
}

